:computer: OrderItem price helper functions refactor
