+++ lib/ex_commerce_web/live/checkout_live/catalogue_item.ex
@@ -432,6 +432,10 @@ defmodule ExCommerceWeb.CheckoutLive.CatalogueItem do
 
   defp assign_nav_title(socket), do: assign(socket, :nav_title, gettext("Back"))
 
+  # ----------------------------------------------------------------------------
+  # Price Getters and Formatters
+  #
+
   defp get_price(price) do
     price
     |> ExCommerceNumeric.format_price()
@@ -439,10 +443,7 @@ defmodule ExCommerceWeb.CheckoutLive.CatalogueItem do
   end
 
   defp get_price(price, price_modifier) do
-    price =
-      price
-      |> ExCommerceNumeric.format_price()
-
+    price = ExCommerceNumeric.format_price(price)
     price_modifier = ExCommerceNumeric.format_price(price_modifier)
 
     CatalogueItemOption.apply_discount(price, price_modifier)
@@ -451,97 +452,9 @@ defmodule ExCommerceWeb.CheckoutLive.CatalogueItem do
 
   defp prepend_currency(price), do: "$#{price}"
 
-  defp get_total_price(%Ecto.Changeset{changes: changes, data: data} = cs) do
-    variant_price = get_variant_price(changes, data)
-    option_groups_price = get_option_groups_price(cs)
-    %{quantity: quantity} = changes
-
-    ExCommerceNumeric.add(variant_price, option_groups_price)
-    |> ExCommerceNumeric.mult(quantity)
-    |> get_price()
-  end
-
-  defp get_variant_price(changes, %OrderItem{variants: variants}) do
-    case Map.get(changes, :variant_id) do
-      nil ->
-        0
-
-      variant_id ->
-        %CatalogueItemVariant{price: price} =
-          Enum.find(variants, fn %CatalogueItemVariant{id: id} ->
-            id == variant_id
-          end)
-
-        price
-    end
-  end
-
-  defp get_option_groups_price(%Ecto.Changeset{
-         changes: %{option_groups: ogs},
-         data: %OrderItem{available_option_groups: aogs}
-       }) do
-    Enum.reduce(
-      aogs.values,
-      ExCommerceNumeric.format_price(0),
-      fn %CatalogueItemOptionGroup{id: id} = og, acc ->
-        get_total_option_group_price(
-          Map.fetch!(ogs, id)["value"],
-          og
-        )
-        |> ExCommerceNumeric.add(acc)
-      end
-    )
-  end
-
-  defp get_option_groups_price(%Ecto.Changeset{}),
-    do: ExCommerceNumeric.format_price(0)
-
-  defp get_total_option_group_price([], %CatalogueItemOptionGroup{}),
-    do: ExCommerceNumeric.format_price(0)
-
-  defp get_total_option_group_price(option_ids, %CatalogueItemOptionGroup{
-         options: options
-       })
-       when is_list(option_ids) do
-    Enum.reduce(
-      options,
-      ExCommerceNumeric.format_price(0),
-      fn %CatalogueItemOption{
-           id: id,
-           catalogue_item_variant: %CatalogueItemVariant{price: price},
-           price_modifier: price_modifier
-         },
-         acc ->
-        case id in option_ids do
-          true ->
-            ExCommerceNumeric.add(
-              acc,
-              CatalogueItemOption.apply_discount(price, price_modifier)
-            )
-
-          false ->
-            acc
-        end
-      end
-    )
-  end
-
-  defp get_total_option_group_price(nil, %CatalogueItemOptionGroup{}),
-    do: ExCommerceNumeric.format_price(0)
-
-  defp get_total_option_group_price(option_id, %CatalogueItemOptionGroup{
-         options: options
-       })
-       when is_binary(option_id) do
-    %CatalogueItemOption{
-      catalogue_item_variant: %CatalogueItemVariant{price: price},
-      price_modifier: price_modifier
-    } =
-      Enum.find(options, 0, fn %CatalogueItemOption{id: id} ->
-        option_id == id
-      end)
-
-    CatalogueItemOption.apply_discount(price, price_modifier)
+  defp get_total_price(%Ecto.Changeset{data: %OrderItem{}} = changeset) do
+    OrderItem.get_total_price(changeset)
+    |> prepend_currency()
   end
 
   # ----------------------------------------------------------------------------
